// wardn/frontend/lib/api/generated/model/mCPServerInstallationToolsResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.1.0
 */
import type { MCPServerInstallationToolsResponseCache } from './mCPServerInstallationToolsResponseCache';
import type { MCPServerToolRead } from './mCPServerToolRead';

export interface MCPServerInstallationToolsResponse {
  cache?: MCPServerInstallationToolsResponseCache;
  configName: string;
  serverName: string;
  serverVersion: string;
  tools: MCPServerToolRead[];
}