// wardn/frontend/lib/api/generated/model/mCPServerInstallationToolValidationResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.1.0
 */
import type { MCPServerInstallationToolValidationResponseResult } from './mCPServerInstallationToolValidationResponseResult';
import type { MCPServerInstallationToolValidationResponseStatus } from './mCPServerInstallationToolValidationResponseStatus';

export interface MCPServerInstallationToolValidationResponse {
  configName: string;
  error?: string;
  isError: boolean;
  result?: MCPServerInstallationToolValidationResponseResult;
  serverName: string;
  status: MCPServerInstallationToolValidationResponseStatus;
  toolName: string;
  validatedAt: string;
}