// wardn/frontend/lib/api/generated/model/mCPServerToolRead.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.1.0
 */
import type { MCPServerToolReadAnnotations } from './mCPServerToolReadAnnotations';
import type { MCPServerToolReadInputSchema } from './mCPServerToolReadInputSchema';
import type { MCPServerToolReadOutputSchema } from './mCPServerToolReadOutputSchema';

export interface MCPServerToolRead {
  annotations?: MCPServerToolReadAnnotations;
  description: string;
  inputSchema: MCPServerToolReadInputSchema;
  outputSchema?: MCPServerToolReadOutputSchema;
  serverName: string;
  serverVersion: string;
  title: string;
  toolName: string;
}